@@ -9,25 +9,32 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@types/jsonwebtoken": "^9.0.6",
     "animate.css": "^4.1.1",
     "aos": "^2.3.4",
     "bootstrap": "^5.3.3",
-    "next": "14.2.2",
-    "react": "^18",
-    "react-bootstrap": "^2.10.2",
-    "react-dom": "^18",
-    "react-icons": "^5.1.0",
-    "react-leaflet": "^4.2.1",
     "crypto": "^1.0.1",
     "date-fns": "^3.6.0",
     "dotenv": "^16.4.5",
     "jsonwebtoken": "^9.0.2",
+    "leaflet": "^1.9.4",
     "mysql2": "^3.9.1",
+    "next": "14.2.2",
     "nodemailer": "^6.9.13",
     "otp-generator": "^4.0.1",
+    "react": "^18.3.1",
+    "react-bootstrap": "^2.10.2",
+    "react-dom": "^18.3.1",
+    "react-icons": "^5.1.0",
+    "react-leaflet": "^4.2.1",
+    "react-otp-input": "^3.1.1",
+    "react-toastify": "^10.0.5",
+    "sass": "^1.77.1",
+    "sweetalert2": "^11.11.0",
     "uuid": "^9.0.1"
   },
   "devDependencies": {
+    "@types/node": "20.12.11",
     "eslint": "^8",
     "eslint-config-next": "14.2.2"
   }
