@@ -3,52 +3,245 @@
 import { useEffect, useState } from "react";
 import Head from "next/head";
 import Link from "next/link";
+import { useRouter } from 'next/navigation';
 import "../../styles/auth.css";
+import { ToastContainer, toast } from 'react-toastify';
+import Swal from 'sweetalert2'
+import 'react-toastify/dist/ReactToastify.css';
 
-const Form = ({ onSubmit }) => {
+const apiUrl = process.env.NEXT_PUBLIC_API_URL
 
-     const [formData, setFormData] = useState({
-          fullName: '',
-          email: '',
-          password: '',
+
+const sendData = async (formData) => {
+  try {
+
+    const res = await fetch(`${apiUrl}/api/signup`, {
+      method: 'POST',
+      body: JSON.stringify(formData),
+      cache: 'no-store',
+      headers: {
+        "Content-Type": "application/json",
+      },
+    })
+    if (!res.ok) {
+
+      throw new Error('Failed to fetch data')
+    }
+
+    const data = await res.json()
+
+    const authToken = data.data
+
+    localStorage.setItem('authToken', authToken);
+
+    sessionStorage.setItem('email', formData.email);
+    console.log(data);
+    return data;
+  } catch (error) {
+    console.error('An error occurred:', error);
+  }
+}
+
+const LoginData = async (formData, rememberMe) => {
+  try {
+    const res = await fetch(`${apiUrl}/api/login`, {
+      method: 'POST',
+      body: JSON.stringify(formData),
+      cache: 'no-store',
+      headers: {
+        "Content-Type": "application/json",
+      },
+    })
+    if (!res.ok) {
+      throw new Error('Failed to fetch data')
+    }
+
+    const data = await res.json()
+    console.log(data);
+
+    const authToken = data.data
+
+    if (rememberMe) {
+      // Store the token in localStorage for long-term persistence
+      localStorage.setItem('authToken', authToken);
+    } else {
+      // Store the token in session storage for short-term persistence
+      sessionStorage.setItem('authToken', authToken);
+    }
+    return data;
+  } catch (error) {
+    console.error('An error occurred:', error);
+  }
+}
+
+console.log(process.env.NEXT_PUBLIC_API_URL);
+
+const Form = () => {
+
+
+  const [passwordError, setPasswordError] = useState('');
+  const [confirmPasswordError, setConfirmPasswordError] = useState('');
+
+  const router = useRouter();
+
+  const [formData, setFormData] = useState({
+    fullName: '',
+    email: '',
+    password: '',
+    confirmPassword: '',
+  });
+
+
+  const handleChange = (e) => {
+    const { name, value } = e.target;
+    setFormData((prev) => ({
+      ...prev,
+      [name]: value,
+    }));
+
+    if (name === 'password') {
+      if (!(value.length >= 6 && /[A-Za-z]/.test(value) && /[0-9]/.test(value))) {
+        setPasswordError('Password must have at least: 6 characters, one letter and one number');
+      } else {
+        setPasswordError('');
+      }
+    }
+
+    // Reset confirm password error if user edits password
+    if (name === 'confirmPassword') {
+
+      if (value !== formData.password) {
+        setConfirmPasswordError('Passwords do not match');
+      } else {
+        setConfirmPasswordError('');
+      }
+
+    }
+
+
+  };
+
+
+  const [rememberMe, setRememberMe] = useState(false);
+
+  const handleCheckboxChange = (e) => {
+    setRememberMe(e.target.checked);
+  };
+
+  const handleSubmitLogin = async (e) => {
+    e.preventDefault();
+
+    try {
+      const response = await LoginData(formData, rememberMe);
+
+      if (response.success === true) {
+        Swal.fire({
+          title: response.message,
+          text: "Welcome to SaveDrive",
+          icon: "success",
+          showConfirmButton: false,
+          timer: 3000
         });
 
-        const [popup, setPopup] = useState({
-          show: false,
-          message: "",
-          type: "", // "success" or "error"
-        });
+        setFormData([])
 
-        const handleShowPopup = (message, type) => {
-          setPopup({ show: true, message, type });
-        };
+        router.push('/enter-address');
+        // toast.success(response.message, {
+        //   position: "top-center"
+        // });
+        // setTimeout(() => {
+        //   router.replace('/about');
+        // }, 1000);
 
-
-        const handleChange = (e) => {
-          const { name, value } = e.target;
-          setFormData((prev) => ({
-            ...prev,
-            [name]: value,
-          }));
-        };
-      
-        const handleSubmit = (e) => {
-          e.preventDefault(); // Prevent default form submission behavior
-          // onSubmit(formData); // Send the data to the parent component or server
-
-          onSubmit(formData)
-          .then((response) => {
-            // Check the response and call `handleShowPopup` based on the result
-            if (response.success === true) {
-              handleShowPopup("Account created successfully!", "success");
-            } else {
-              handleShowPopup("An error occurred. Please try again.", "error");
-            }
-          })
-          .catch((error) => {
-            handleShowPopup("An error occurred. Please try again.", "error");
+      } else if (response.success === false) {
+        if (response.message === "Please verify your email") {
+          Swal.fire({
+            icon: "info",
+            title: "One more step",
+            text: response.message,
+            showConfirmButton: false,
+            timer: 2000
           });
-        };
+
+          router.replace('/otp');
+        } else {
+          Swal.fire({
+            icon: "error",
+            title: "Oops...",
+            text: response.message,
+            showConfirmButton: false,
+            timer: 2000
+          });
+        }
+        
+
+        // toast.error(response.message, {
+        //   position: "top-center"
+        // });
+      }
+    } catch (error) {
+      Swal.fire({
+        icon: "error",
+        title: "Oops...",
+        text: "Something went wrong!",
+
+      });
+    }
+  };
+
+
+  const handleSubmitSignup = async (e) => {
+    e.preventDefault();
+
+    try {
+
+      // Check if password and confirm password match
+      if (formData.password !== formData.confirmPassword) {
+        // toast.error('Passwords do not match');
+        setConfirmPasswordError('Passwords do not match');
+        return;
+      } else {
+        setConfirmPasswordError('');
+      }
+      const response = await sendData(formData);
+
+      if (response.success === true) {
+
+        Swal.fire({
+          title: response.message,
+          text: "A One-Time Verification Pin has been sent to your email. Use it to verify your email.",
+          icon: "success",
+          showConfirmButton: false,
+          timer: 3000
+        });
+        setFormData([])
+
+
+        setTimeout(() => {
+          router.replace('/otp');
+        }, 1000);
+
+      } else if (response.success === false) {
+        Swal.fire({
+          icon: "error",
+          title: "Oops...",
+          text: response.message,
+          showConfirmButton: false,
+          timer: 2000
+        });
+      }
+    } catch (error) {
+      Swal.fire({
+        icon: "error",
+        title: "Oops...",
+        text: "Something went wrong!",
+        footer: '<a href="#">Why do I have this issue?</a>'
+      });
+    }
+  };
+
+
+
 
 
   useEffect(() => {
@@ -86,7 +279,7 @@ const Form = ({ onSubmit }) => {
           />
         </Head>
 
-        <div className="wrapper">
+        <div className="wrapper shadow-lg">
           <div className="title-text">
             <div className="title login">
               Welcome Back
@@ -112,15 +305,22 @@ const Form = ({ onSubmit }) => {
             <div className="form-inner">
               <form className="login" >
                 <div className="field mt-4">
-                  <input type="text" placeholder="Email Address" required />
+                  <input type="text"
+                    name="email"
+                    value={formData.email}
+                    onChange={handleChange} placeholder="Email Address" required />
                 </div>
                 <div className="field  mt-4">
-                  <input type="password" placeholder="Password" required />
+                  <input type="password" name="password"
+                    value={formData.password}
+                    onChange={handleChange} placeholder="Password" required />
                 </div>
-               
+
                 <div className="d-flex justify-content-between align-items-center mt-2 pass-link">
                   <div className="d-flex align-items-center ms-2">
-                    <input type="checkbox" id="rememberMe" className="me-1" />
+                    <input type="checkbox"
+                      checked={rememberMe}
+                      onChange={handleCheckboxChange} id="rememberMe" className="me-1" />
                     <label htmlFor="rememberMe" className="">
                       Remember Me
                     </label>
@@ -131,7 +331,7 @@ const Form = ({ onSubmit }) => {
                 </div>
                 <div className="field btns mt-4">
                   <div className="btn-layer"></div>
-                  <input type="submit" value="Login" />
+                  <input type="submit" value="Login" onClick={handleSubmitLogin} />
                 </div>
                 <div className="signup-link mt-2">
                   Not a member? <Link href="" className="links-color">Signup now</Link>
@@ -140,32 +340,37 @@ const Form = ({ onSubmit }) => {
               <form className="signup" >
                 <div className="field">
                   <input type="text"
-                  name="fullName"
-                  value={formData.fullName}
-                  onChange={handleChange} placeholder="Fullname" required />
+                    name="fullName"
+                    value={formData.fullName}
+                    onChange={handleChange} placeholder="Fullname" required />
                 </div>
                 <div className="field">
                   <input type="email"
-                   name="email"
-                   value={formData.email}
-                   onChange={handleChange} placeholder="Email Address" required />
+                    name="email"
+                    value={formData.email}
+                    onChange={handleChange} placeholder="Email Address" required />
                 </div>
                 <div className="field">
                   <input type="password"
-                  name="password"
-                  value={formData.password}
-                  onChange={handleChange} placeholder="Password" required />
+                    name="password"
+                    value={formData.password}
+                    onChange={handleChange} placeholder="Password" required />
                 </div>
+                {passwordError && <p className="error">{passwordError}</p>}
                 <div class="field">
                   <input
                     type="password"
+                    name="confirmPassword"
+                    value={formData.confirmPassword}
+                    onChange={handleChange}
                     placeholder="Confirm password"
                     required
                   />
+                  {confirmPasswordError && <p className="error">{confirmPasswordError}</p>}
                 </div>
                 <div className="field btns">
                   <div className="btn-layer"></div>
-                  <input type="submit" value="Signup" onClick={handleSubmit}/>
+                  <input type="submit" value="Signup" onClick={handleSubmitSignup} />
                 </div>
                 <div className="signup-link mt-2">
                   Already have an account?{" "}
@@ -174,14 +379,6 @@ const Form = ({ onSubmit }) => {
                   </Link>
                 </div>
               </form>
-
-                  {/* Conditional rendering for popup */}
-      {popup.show && (
-        <div className={`popup ${popup.type}`}>
-          <span>{popup.message}</span>
-          <button onClick={() => setPopup({ ...popup, show: false })}>Close</button>
-        </div>
-      )}
             </div>
           </div>
         </div>
