@@ -5,18 +5,20 @@ import Link from "next/link";
 import LoggedInBanner from "../components/loggedInHero";
 import AvailableDrivers from "./components/available-drivers";
 import Footer from "../components/footer";
+import LoggedInNav from "../components/loggedInNav";
 
 
 const RideOptions = () => {
 
   return (
     <>
-    <div className=" bg-body-secondary">
-    <LoggedInBanner question="Ready for your ride?" command="Choose the best option for you" />
+  
+      <LoggedInNav />
+    {/* <LoggedInBanner question="Ready for your ride?" command="Choose the best option for you" /> */}
    <AvailableDrivers />
 
-   <Footer />
-    </div>
+   {/* <Footer /> */}
+
   
     </>
   );
