@@ -5,10 +5,14 @@ import './styles/globals.css'
 import './styles/default.css'
 import './styles/fontawesome.css'
 import BootstrapJs from "./components/bootstrapJS";
+import 'react-toastify/dist/ReactToastify.css';
+import { ToastContainer, toast } from 'react-toastify';
+// import "https://unpkg.com/leaflet@1.9.4/dist/leaflet.css"
+import "leaflet/dist/leaflet.css"
+import Swal from 'sweetalert2'
 
 
 const inter = Inter({ subsets: ["latin"] });
-
 export const metadata = {
   title: "Create Next App",
   description: "Generated by create next app",
@@ -19,6 +23,7 @@ export default function RootLayout({ children }) {
     <html lang="en">
       <body className={inter.className}>{children}
       <BootstrapJs />
+      <ToastContainer />
       </body>
     </html>
   );
